refactored legend icon separately from treeNode
